// web/src/app/models/equiv_course_joined.ts
export class EquivCourseJoined {
  EquivID: number;
  Status: string;
  LocalCourseID: number;
  LocalCourseName: string
  ForeignCourseID: number;
  ForeignCourseName: string;
  SchoolName: string;
};
//{"EquivID":1,"Status":"accepted","LocalCourseID":9,"LocalCourseName":"COEN 210 - Computer Architecture","ForeignCourseID":1,"ForeignCourseName":"CMPE 200 - Computer Architecture","SchoolName":"San Jose State University"}