// web/src/app/models/local_course_joined.ts
export class LocalCourseJoined {
	EquivID: number;
	Status: string;
	LocalCourseID: number;
	LocalCourseName: string;
	ForeignCourseID: number;
	ForeignCourseName: string;
	SchoolName: string;
};
//{"EquivID":null,"Status":null,"LocalCourseID":1,"LocalCourseName":"COEN 200 - Logic Analysis & Synthesis","ForeignCourseID":null,"ForeignCourseName":null,"SchoolName":null},