// web/src/app/models/change_joined.ts
export class ChangeJoined {
	LocalCourseName: string;
	ForeignCourseName: string;
	SchoolName: string;
	NewStatus: string;
	Notes: string;
	Date: string; //is there a date type?
	UserName: string;
	UserEmail: string;
};
//{"LocalCourseName":"COEN 233 - Computer Networks","ForeignCourseName":"CSC 745 - Computer Networks","SchoolName":"San Francisco State University","NewStatus":"accepted","Notes":"added via recommender system","Date":"2017-10-11 21:11:53","UserName":"Andrew Leonard","UserEmail":"aleonard@scu.edu"}